+++ src/Maestro/Maestro.MergePolicies/MergePolicyEvaluationResult.cs
@@ -1,0 +1,14 @@
+namespace Maestro.MergePolicies
+{
+    public class MergePolicyEvaluationResult
+    {
+        internal MergePolicyEvaluationResult(bool succeeded, string message)
+        {
+            Succeeded = succeeded;
+            Message = message;
+        }
+
+        public bool Succeeded { get; }
+        public string Message { get; }
+    }
+}
